+++ src/ElasticSan/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for private endpoint for ElasticSan
 
 ## Version 0.1.0
 * First preview release for module Az.ElasticSan